// src/api/nexuscore/model/pinOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Nexuscore
 * The backend for all Everthorn internal services
 * OpenAPI spec version: 1.0.0
 */

export interface PinOut {
  /** The pin's ID */
  id: number;
  /** The name of the pin */
  name: string;
  /** A short description of the pin, such as what the shop sells, etc. */
  description: string;
  /** The coordinates of the pin */
  coordinates: number[];
  /** The dimension of the pin */
  dimension: string;
  /** The type of pin this is */
  pin_type: string;
}