// src/api/nexuscore/model/projectOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Nexuscore
 * The backend for all Everthorn internal services
 * OpenAPI spec version: 1.0.0
 */
import type { StatusEnum } from './statusEnum.ts';
import type { UserOut } from './userOut.ts';

export interface ProjectOut {
  /** The string ID of the project */
  project_id: string;
  /** The name of the project */
  name: string;
  thread_id: number | null;
  /** The coordinates of the project */
  coordinates: number[];
  /** A short description of the project */
  description: string;
  completed_on: string | null;
  pin_id: number | null;
  /** The dimension of the project */
  dimension: string;
  started_on: string | null;
  /** The project status */
  status: StatusEnum;
  /** When the status was last updated */
  status_since: string;
  /** The owner of the project, in the form of a User object */
  owner: UserOut;
}